+++ mecs.py
@@ -224,8 +224,8 @@ class CommandBuffer():
         Associate the buffer with an entity storage.
         """
 
-        self.storage = storage
-        self.commands = []
+        self._storage = storage
+        self._commands = []
 
     def __enter__(self):
         return self
@@ -233,7 +233,11 @@ class CommandBuffer():
     def __exit__(self, type, value, traceback):
         self.flush()
 
-    def create(self, component=None):
+    @property
+    def storage(self):
+        return self._storage
+
+    def defer_create(self, component=None):
         """
         Store a future call to the *create()* method of the associated storage.
 
@@ -244,32 +248,43 @@ class CommandBuffer():
         """
 
         entity = _generate_new_entity_id()
-        self.commands.append((self.storage.create, (component, entity)))
+        self._commands.append((self._storage.create, (component, entity)))
         
-    def set(self, entity, component):
+    def defer_destroy(self, entity):
+        """
+        Store a future call to the *destroy()* method of the associated storage.
+
+        *New in version 1.3.*
+        """
+
+        self._commands.append((self._storage.destroy, (entity,)))
+
+    def defer_set(self, entity, component):
         """
         Store a future call to the *set()* method of the associated storage.
 
         *New in version 1.2.*
         """
 
-        self.commands.append((self.storage.set, (entity, component)))
+        self._commands.append((self._storage.set, (entity, component)))
 
-    def delete(self, entity, component_type):
+    def defer_delete(self, entity, component_type):
         """
         Store a future call to the *delete()* method of the associated storage.
 
         *Changed in version 1.2:* Added support for multiple component types.
         """
 
-        self.commands.append((self.storage.delete, (entity, component_type)))
+        self._commands.append((self._storage.delete, (entity, component_type)))
 
-    def destroy(self, entity):
+    def defer_clear(self):
         """
-        Store a future call to the *destroy()* method of the associated storage.
+        Storage a future call to the *clear()* method of the accociated storage.
+
+        *New in version 1.3.*
         """
 
-        self.commands.append((self.storage.destroy, (entity,)))
+        self._commands.append((self._storage.clear,))
 
     def flush(self):
         """
@@ -278,9 +293,20 @@ class CommandBuffer():
         This will apply all commands that have previously been stored in the buffer to the associated storage. Note that if any of the arguments in these commands were invalid, exceptions may arrise.
         """
 
-        for cmd, args in self.commands:
+        for cmd, args in self._commands:
             cmd(*args)
-        self.commands.clear()
+        self.clear()
+
+    def clear(self):
+        """
+        Clear the buffer.
+
+        This does not apply any of the stored commands.
+
+        *New in version 1.3.*
+        """
+
+        self._commands.clear()
 
 class Storage():
     """